How Much Does Daytona Cost problem is solved of rathing 800
simple approach, just observed that we have to take subsegment in which the given element is most common and one good thing I had observed that we have to take subsegment by our own and there is no restrictions of selecting subsegment. Additionally observed from one test case in which arr is of size 2 and the subsegment is of length is one. so this will print No if and only if the given element is not present in arr. so for that firstly i had sorted arr using merge sort and then used binary search. Simple.